+++ Editor/Scripts/Typed/FloatGameEventInspector.cs
@@ -1,0 +1,9 @@
+﻿using UnityEditor;
+
+namespace FredericRP.EventManagement
+{
+  [CustomEditor(typeof(FloatGameEvent))]
+  public class FloatGameEventInspector : OneTypeGameEventInspector<float>
+  {
+  }
+}