+++ Editor/Scripts/Typed/IntGameEventInspector.cs
@@ -1,0 +1,9 @@
+using UnityEditor;
+
+namespace FredericRP.EventManagement
+{
+  [CustomEditor(typeof(IntGameEvent))]
+  public class IntGameEventInspector : OneTypeGameEventInspector<int>
+  {
+  }
+}