+++ Editor/Scripts/Typed/StringGameEventInspector.cs
@@ -1,0 +1,9 @@
+using UnityEditor;
+
+namespace FredericRP.EventManagement
+{
+  [CustomEditor(typeof(StringGameEvent))]
+  public class StringGameEventInspector : OneTypeGameEventInspector<string>
+  {
+  }
+}