+++ Runtime/Scripts/TypedGameEvent/Typed/FloatGameEvent.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/Float Game Event")]
+  public class FloatGameEvent : OneTypeGameEvent<float>
+  {
+  }
+}