+++ Runtime/Scripts/TypedGameEvent/Typed/IntGameEvent.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/Integer Game Event")]
+  public class IntGameEvent : OneTypeGameEvent<int>
+  {
+  }
+}