+++ Runtime/Scripts/TypedGameEvent/Typed/StringGameEvent.cs
@@ -1,0 +1,9 @@
+using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/String Game Event")]
+  public class StringGameEvent : OneTypeGameEvent<string>
+  {
+  }
+}