+++ Runtime/Scripts/ValueReference/Typed/StringReference.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+
+namespace FredericRP.EventManagement
+{
+  [CreateAssetMenu(menuName = "FredericRP/Value Reference/String")]
+  public class StringReference : GenericReference<string>
+  {
+  }
+}